requirements.txt
streamlit
pandas
numpy
scikit-learn==1.5.2
xgboost==1.7.6
tensorflow
sentence-transformers
faiss-cpu
langchain
langchain-community
shap==0.49.1
matplotlib
plotly
joblib
sumy
groq
python-dotenv
reportlab
seaborn
imbalanced-learn